Rename geometry package to math
