+++ CMakeLists.txt
@@ -9,8 +9,8 @@ project(titanforge CXX)
 
 add_subdirectory(packages/common)
 add_subdirectory(packages/common_test)
-add_subdirectory(packages/geometry)
-add_subdirectory(packages/geometry_test)
 add_subdirectory(packages/graphics)
 add_subdirectory(packages/graphics_test)
+add_subdirectory(packages/math)
+add_subdirectory(packages/math_test)
 add_subdirectory(packages/sampleapp)
