--- packages/geometry_test/README.md
@@ -1,5 +1,0 @@
-# Geometry Test
-
-This code contains unit tests for `geometry` module code.
-
-For instructions on how to build and run this code, please refer to the [parent README](../../README.md).
