--- packages/geometry_test/src/main.cpp
@@ -1,2 +1,0 @@
-#define BOOST_TEST_MODULE GeometryTest
-#include <boost/test/unit_test.hpp>
