+++ packages/graphics/README.md
@@ -1,5 +1,5 @@
 # Graphics
 
-This library contains code for rendering 2D and 3D geometry using the OpenGL standard.
+This library contains code for rendering 2D and 3D objects using the OpenGL standard.
 
 For instructions on how to build and run this code, please refer to the [parent README](../../README.md).
