+++ packages/graphics_test/src/shaders/ShaderTest.cpp
@@ -5,7 +5,7 @@
 #include <graphics/shaders/Shader.h>
 #include <graphics/shaders/ShaderManager.h>
 #include <graphics/shaders/glsl/Basic.h>
-#include <geometry/Matrix4.h>
+#include <math/Matrix4.h>
 #include <common/exceptions/IllegalArgumentException.h>
 #include <common/exceptions/InstantiationException.h>
 #include <glad/glad.h>
