+++ packages/math/README.md
@@ -1,0 +1,5 @@
+# Math
+
+This library contains reusable numerical computations, vector and matrix operations, and mathematical functions.
+
+For instructions on how to build and run this code, please refer to the [parent README](../../README.md).
