+++ packages/math/src/Matrix4.cpp
@@ -1,6 +1,6 @@
-#include <geometry/Matrix4.h>
-#include <geometry/Matrix3.h>
-#include <geometry/Vector3.h>
+#include <math/Matrix4.h>
+#include <math/Matrix3.h>
+#include <math/Vector3.h>
 #include <common/Utils.h>
 #include <common/exceptions/IllegalArgumentException.h>
 #include <string>
