+++ packages/math_test/CMakeLists.txt
@@ -1,4 +1,4 @@
-project(GeometryTest CXX)
+project(MathTest CXX)
 
 set(CMAKE_BUILD_TYPE Debug)
 if(CMAKE_SYSTEM_NAME STREQUAL "Linux")
@@ -10,7 +10,7 @@ enable_testing()
 find_package(Boost REQUIRED COMPONENTS unit_test_framework)
 
 # Executable files
-add_executable(GeometryTest
+add_executable(MathTest
     src/main.cpp
     src/Matrix3Test.cpp
     src/Matrix4Test.cpp
@@ -19,13 +19,13 @@ add_executable(GeometryTest
 )
 
 # Linked libraries
-target_link_libraries(GeometryTest
+target_link_libraries(MathTest
     PUBLIC Boost::unit_test_framework
     PUBLIC Common::Common
-    PUBLIC Geometry::Geometry)
+    PUBLIC Math::Math)
 
 # Outputs
-set_target_properties(GeometryTest PROPERTIES
+set_target_properties(MathTest PROPERTIES
     RUNTIME_OUTPUT_DIRECTORY "${CMAKE_CURRENT_BINARY_DIR}/bin"
     RUNTIME_OUTPUT_DIRECTORY_DEBUG "${CMAKE_CURRENT_BINARY_DIR}/bin"
     RUNTIME_OUTPUT_DIRECTORY_RELEASE "${CMAKE_CURRENT_BINARY_DIR}/bin"
