+++ packages/math_test/src/Matrix3Test.cpp
@@ -1,6 +1,6 @@
 #include <boost/test/unit_test.hpp>
-#include <geometry/Matrix3.h>
-#include <geometry/Vector3.h>
+#include <math/Matrix3.h>
+#include <math/Vector3.h>
 #include <common/Utils.h>
 #include <common/Constants.h>
