+++ packages/math_test/src/main.cpp
@@ -1,0 +1,2 @@
+#define BOOST_TEST_MODULE MathTest
+#include <boost/test/unit_test.hpp>
