+++ packages/sampleapp/CMakeLists.txt
@@ -7,7 +7,7 @@ add_executable(SampleApp
 # Linked libraries
 target_link_libraries(SampleApp
     PRIVATE Common::Common
-    PRIVATE Geometry::Geometry
+    PRIVATE Math::Math
     PRIVATE Graphics::Graphics)
 
 # Outputs
