+++ packages/sampleapp/src/main.cpp
@@ -9,8 +9,8 @@
 #include <graphics/shaders/Shader.h>
 #include <graphics/shaders/ShaderManager.h>
 #include <graphics/objects/Mesh.h>
-#include <geometry/Vector3.h>
-#include <geometry/Matrix4.h>
+#include <math/Vector3.h>
+#include <math/Matrix4.h>
 #include <common/Utils.h>
 
 /**
